fix(rulers): reduce code duplication for linear rulers
